Start of Game
With this commit and blessing of almighty, I now on this planet earth
would like to kickstart a learning project or process or repository or
whatsoever you would want to name it. This tauri build is a learning
step which with success in future helps me to convert many node based
backends into rust based backends in samagra.

Signed-off-by: Harshil Jani <harshiljani2002@gmail.com>