Add Android Release
Open source communities are heaven. So, It was around 2 hrs I was
busting my face around a github action that could help me building the
tauri Android APK. Then , I decided to text in the tauri discord channel
and the guy replied back within 2 mins with the action which worked.
This is power of open source and we should not simply underestimate this
beauty.

Signed-off-by: Harshil Jani <harshiljani2002@gmail.com>